+++ src/mfdrv.h
@@ -35,9 +35,9 @@ void mf_drawpixel(const Canvas *canvas, const VPoint *vp);
 void mf_drawpolyline(const Canvas *canvas, const VPoint *vps, int n, int mode);
 void mf_fillpolygon(const Canvas *canvas, const VPoint *vps, int nc);
 void mf_drawarc(const Canvas *canvas,
-    const VPoint *vp1, const VPoint *vp2, int a1, int a2);
+    const VPoint *vp1, const VPoint *vp2, double a1, double a2);
 void mf_fillarc(const Canvas *canvas,
-    const VPoint *vp1, const VPoint *vp2, int a1, int a2, int mode);
+    const VPoint *vp1, const VPoint *vp2, double a1, double a2, int mode);
 void mf_putpixmap(const Canvas *canvas,
     const VPoint *vp, int width, int height, char *databits,
     int pixmap_bpp, int bitmap_pad, int pixmap_type);
